# docs/integration/openapi-iteration3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "教育服务系统 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "基础设施",
          "基础设施"
        ],
        "summary": "用户登录",
        "description": "用户登录接口。\n\n对应 API 文档第 4.2 节。\n\n流程:\n  1. 验证用户名 + 密码（bcrypt 校验）\n  2. 签发 JWT Token（24h 有效期）\n  3. 返回 Token + 用户信息\n\n请求体:\n  {\"username\": \"admin\", \"password\": \"123456\"}\n\n成功响应:\n  {code: 0, message: \"success\", data: {user_id, username, real_name, user_type, access_token, ...}}\n\n错误响应:\n  40103: 用户名或密码错误\n  40103: 账号已被禁用",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "基础设施",
          "基础设施"
        ],
        "summary": "当前用户信息",
        "description": "获取当前登录用户的详细信息。\n\n对应 API 文档第 4.3 节。\n\n请求头:\n  Authorization: Bearer {access_token}\n\n成功响应:\n  {code: 0, data: {user_id, username, real_name, user_type, department, contact_info, ...}}",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "description": "登录账号",
            "examples": [
              "admin"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "description": "登录密码（明文，传输时应走 HTTPS）",
            "examples": [
              "123456"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "用户登录请求体。\n\n对应 API 文档第 4.2 节 POST /api/v1/auth/login。\n前端提交用户名和明文密码，后端用 bcrypt 校验。\n\n校验规则:\n  - username: 必填，1-64 字符\n  - password: 必填，1-128 字符（仅做长度校验，不对密码复杂度做强限制）"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "DifyServiceToken": {
        "type": "http",
        "description": "Dify HTTP 节点调用 FastAPI 白名单接口的服务令牌",
        "scheme": "bearer"
      }
    }
  }
}